Improve error messages with Latexmk is not found.
Not being an atom expert, I ran into some trouble getting the latex
plugin to build my document. It was very hard for me to figure out from
the error messages that I simply needed to update the Tex Path setting
of the plugin. With this modification, users should be able to resolve
the same problem faster.

I kept the modification as simple as possible so it will need updating
as more builders are added.