Add a Skim.app opener
Only adds the actual opener, with no code to make it accessible to the user.
This would probably be done using some kind of option, but I'm not sure what
the best implementation of PDF viewer selection options would be.

This is kind of in preparation of introducing SyncTeX capabilities, since
Preview.app doesn't support this (?). Skim.app does, through a utility
included in its application contents, but using that utility requires that
the caller knows the full path to Skim (standard is `/Applications/Skim.app`)
which may be something to have in ming when designing a PDF viewer option.