Ensure we only attempt to build supported files
To ensure we only attempt to build supported filetypes, a
naïve check has been implemented that simply checks the file extension;
right now unless the extension is .tex or .lhs the build invokation will
be aborted, and an info message will be logged to the Atom dev console.

This should hopefully resolve some issues/annoyances, but since the
supported extensions are hardcoded right now, new annoyances are likely
to surface. Perhaps extensions should be configurable?