Handle missing output file info in log files
This is an attempt to fix #46 (and related issues) which are caused
by "Output written on .." text missing in log files. This caused
problems due to `outputFilePath` being set to `null` yet the status code
was 0, thus the build was treated as a success, but paths ended up
with an invalid "null" segment.

Added a regression style test that should (hopefully) ensure this
problem does not get re-introduced.