Improve TeX path resolution, and test coverage
Changes the old behavior of only using the $PATH environment variable
in the latexmk child process' PATH if the resolved TeX path contained
the substitution marker `$PATH`. Substitutions are still supported,
but the new default is to not use it, and when it is not present, the
inherited $PATH environment variable is appended to the configured
TeX path instead.

Refs #41